Features/v3.1/wes workflow compatibility

+++ src/api/mvc/variants.go
@@ -251,6 +251,19 @@ func VariantsIngest(c echo.Context) error {
 				}
 				// -----
 
+				// --- tabix generation
+				tabixFileDir, tabixFileName, tabixErr := ingestionService.GenerateTabix(tmpDestinationFileName)
+				if tabixErr != nil {
+					msg := "Something went wrong: Tabix problem " + gzippedFileName
+					fmt.Println(msg)
+
+					reqStat.State = ingest.Error
+					reqStat.Message = msg
+					ingestionService.IngestRequestChan <- reqStat
+
+					return
+				}
+
 				// ---   push compressed to DRS
 				drsFileId := ingestionService.UploadVcfGzToDrs(cfg.Drs.BridgeDirectory, gzippedFileName, drsUrl, drsUsername, drsPassword)
 				if drsFileId == "" {
@@ -264,9 +277,33 @@ func VariantsIngest(c echo.Context) error {
 					return
 				}
 
-				// ---   remove temporary file now that it has been ingested successfully into DRS
+				// -- push tabix to DRS
+				drsTabixFileId := ingestionService.UploadVcfGzToDrs(cfg.Drs.BridgeDirectory, tabixFileName, drsUrl, drsUsername, drsPassword)
+				if drsTabixFileId == "" {
+					msg := "Something went wrong: DRS Tabix File Id is empty for " + tabixFileName
+					fmt.Println(msg)
+
+					reqStat.State = ingest.Error
+					reqStat.Message = msg
+					ingestionService.IngestRequestChan <- reqStat
+
+					return
+				}
+
+				// ---   remove temporary files now that they have been ingested successfully into DRS
 				if tmpFileRemovalErr := os.Remove(tmpDestinationFileName); tmpFileRemovalErr != nil {
 					msg := fmt.Sprintf("Something went wrong: trying to remove temporary file at %s : %s\n", tmpDestinationFileName, tmpFileRemovalErr)
+					fmt.Println(msg)
+
+					reqStat.State = ingest.Error
+					reqStat.Message = msg
+					ingestionService.IngestRequestChan <- reqStat
+
+					return
+				}
+				tmpTabixFilePath := fmt.Sprintf("%s%s", tabixFileDir, tabixFileName)
+				if tmpTabixFileRemovalErr := os.Remove(tmpTabixFilePath); tmpTabixFileRemovalErr != nil {
+					msg := fmt.Sprintf("Something went wrong: trying to remove temporary file at %s : %s\n", tmpTabixFilePath, tmpTabixFileRemovalErr)
 					fmt.Println(msg)
 
 					reqStat.State = ingest.Error
@@ -414,6 +451,19 @@ func executeGetByIds(c echo.Context, ids []string, isVariantIdQuery bool) error 
 	var es, chromosome, lowerBound, upperBound, reference, alternative, genotype, assemblyId = retrieveCommonElements(c)
 
 	// retrieve other query parameters relevent to this 'get' query ---
+	getSampleIdsOnlyQP := c.QueryParam("getSampleIdsOnly")
+	var (
+		getSampleIdsOnly bool = false
+		getSioErr        error
+	)
+	// only respond sampleIds-only
+	if isVariantIdQuery && len(getSampleIdsOnlyQP) > 0 {
+		getSampleIdsOnly, getSioErr = strconv.ParseBool(getSampleIdsOnlyQP)
+		if getSioErr != nil {
+			log.Fatal(getSioErr)
+		}
+	}
+
 	sizeQP := c.QueryParam("size")
 	var (
 		defaultSize = 100
@@ -475,7 +525,8 @@ func executeGetByIds(c echo.Context, ids []string, isVariantIdQuery bool) error 
 					_id, "", // note : "" is for sampleId
 					reference, alternative,
 					size, sortByPosition,
-					includeInfoInResultSet, genotype, assemblyId)
+					includeInfoInResultSet, genotype, assemblyId,
+					getSampleIdsOnly)
 			} else {
 				// implied sampleId query
 				variantRespDataModel.SampleId = _id
@@ -487,7 +538,8 @@ func executeGetByIds(c echo.Context, ids []string, isVariantIdQuery bool) error 
 					"", _id, // note : "" is for variantId
 					reference, alternative,
 					size, sortByPosition,
-					includeInfoInResultSet, genotype, assemblyId)
+					includeInfoInResultSet, genotype, assemblyId,
+					false)
 			}
 			if searchErr != nil {
 				errorMux.Lock()
@@ -498,28 +550,49 @@ func executeGetByIds(c echo.Context, ids []string, isVariantIdQuery bool) error 
 
 			// query for each id
 
-			docsHits := docs["hits"].(map[string]interface{})["hits"]
-			allDocHits := []map[string]interface{}{}
-			mapstructure.Decode(docsHits, &allDocHits)
+			if !getSampleIdsOnly {
+				docsHits := docs["hits"].(map[string]interface{})["hits"]
+				allDocHits := []map[string]interface{}{}
+				mapstructure.Decode(docsHits, &allDocHits)
 
-			// grab _source for each hit
-			var allSources []models.Variant
+				// grab _source for each hit
+				var allSources []models.Variant
 
-			for _, r := range allDocHits {
-				source := r["_source"].(map[string]interface{})
+				for _, r := range allDocHits {
+					source := r["_source"].(map[string]interface{})
 
-				// cast map[string]interface{} to struct
-				var resultingVariant models.Variant
-				mapstructure.Decode(source, &resultingVariant)
+					// cast map[string]interface{} to struct
+					var resultingVariant models.Variant
+					mapstructure.Decode(source, &resultingVariant)
 
-				// accumulate structs
-				allSources = append(allSources, resultingVariant)
+					// accumulate structs
+					allSources = append(allSources, resultingVariant)
+				}
+
+				fmt.Printf("Found %d docs!\n", len(allSources))
+
+				variantRespDataModel.Count = len(allSources)
+				variantRespDataModel.Results = allSources
+			} else {
+				// TODO: refactor this 'else' statement
+				docsBuckets := docs["aggregations"].(map[string]interface{})["sampleIds"].(map[string]interface{})["buckets"]
+				allDocBuckets := []map[string]interface{}{}
+				mapstructure.Decode(docsBuckets, &allDocBuckets)
+
+				var allSampleIdsOnly []string
+
+				for _, r := range allDocBuckets {
+					sampleId := r["key"].(string)
+
+					// accumulate sample Id's
+					allSampleIdsOnly = append(allSampleIdsOnly, sampleId)
+				}
+
+				fmt.Printf("Found %d docs!\n", len(allSampleIdsOnly))
+
+				variantRespDataModel.Count = len(allSampleIdsOnly)
+				variantRespDataModel.SampleIds = allSampleIdsOnly
 			}
-
-			fmt.Printf("Found %d docs!\n", len(allSources))
-
-			variantRespDataModel.Count = len(allSources)
-			variantRespDataModel.Results = allSources
 
 			respDTOMux.Lock()
 			respDTO.Data = append(respDTO.Data, variantRespDataModel)
